Test with fp64.

+++ nqs/network.py
@@ -57,7 +57,10 @@ def MaskedConv1d(
         )
         bias_shape = [out_chan if c == "C" else 1 for c in out_spec]
         bias_shape = tuple(itertools.dropwhile(lambda x: x == 1, bias_shape))
-        W, b = W_init(rng, kernel_shape), b_init(rng, bias_shape)
+        W, b = (
+            W_init(rng, kernel_shape, dtype=np.float64),
+            b_init(rng, bias_shape, dtype=np.float64),
+        )
         return output_shape, (W, b)
 
     def apply_fun(params, inputs, **kwargs):
